+++ lib/slkecho/option_parser.rb
@@ -16,6 +16,7 @@ module Slkecho
         o.on("--icon-url ICON_URL", "Set user icon image for message by URL.")
         o.on("--icon-emoji ICON_EMOJI", "Set user image for message by emoji.")
         o.on("--message-as-blocks", "Post message as blocks.")
+        o.on("--token TOKEN", "Slack API token.")
       end
     end
 
@@ -29,7 +30,8 @@ module Slkecho
     def build_options(argv)
       option_values = {}
       argv = option_parser.parse(argv, into: option_values)
-      option_values = option_values.transform_keys { _1.to_s.tr("-", "_") }
+      option_values = option_values.transform_keys { _1.to_s.tr("-", "_").to_sym }
+      option_values[:token] ||= ENV.fetch("SLACK_API_TOKEN", nil)
 
       Slkecho::Options.new(option_values).tap do |opt|
         opt.message = fetch_message(argv)