+++ lib/slkecho/options.rb
@@ -15,9 +15,11 @@ module Slkecho
     attribute :icon_url, :string
     attribute :icon_emoji, :string
     attribute :message_as_blocks, :boolean
+    attribute :token, :string
 
     validates :channel, presence: { message: "is required." }, unless: -> { configure }
     validates :message, presence: { message: "is missing." }, unless: -> { configure || !message.nil? }
+    validates :token, presence: { message: "is required." }, unless: -> { configure }
 
     def error_message
       errors.full_messages.join(" ").downcase