+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - feat: add `--configure` option
+- feat: add `--token` option
 
 ## [v2.1.6] - 2025-01-05